[FIC]:Error Eslint

+++ src/services/Business.service.js
@@ -99,7 +99,7 @@ class BusinessService extends BaseService {
       throw error;
     }
     const urlLogo = `${id}/${filename}`;
-    await CloudStorage.deleteImage(businessExists.logo);
+    if (businessExists.logo) await CloudStorage.deleteImage(businessExists.logo);
     await CloudStorage.saveImage(filename, urlLogo);
     await _businessRepository.update(id, {
       logo: urlLogo,
